requirements.txt
docker
gnsq
gevent
pyserial
flask
flask-restful
pymemcache
